Remove haiku similarity query page
While on its face this might seem like the most important analysis
writeup to add, in reality, it's too fragmented, and there were no real
conclusions that I was able to draw from the analysis. It's mostly just
spamming code for code's sake. So remove the damn thing and forget about
it.