Use named volumes to make certs persistent
This means that the acme-companion container does not have to regenerate
the certificates every time it starts up. This makes it less likely to
hit the letsencrypt certificate rate limiting (five new certs a week).

Additionally, use the official, maintained, images
nginxproxy/nginx-proxy, and nginxproxy/acme-companion. These images
originated from the ones that I was originally using, but are now owned
and maintained by the same organization.

Closes #91
Closes #96
Closes #97