Add published and modified date metadata to pages
Use dcterms.available and dcterms.modified as defined by [1] and [2]
respectively.

This metadata will be consumed by a, as-of-yet WIP RSS feed generator.
Draft pages will get a published date of "DRAFT", and index pages will
not get the date metadata. The feed generator will ignore pages with a
DRAFT publish date, as well as pages with no publish date at all.

[1] https://www.dublincore.org/specifications/dublin-core/dcmi-terms/#http://purl.org/dc/terms/available
[2] https://www.dublincore.org/specifications/dublin-core/dcmi-terms/#http://purl.org/dc/terms/modified