Fix haiku generation query parameters being dropped
[1] introduced a bug where the HTTP query parameters were not passed
through the proxy to the research API container. According to [2], if
the proxy_pass is specified with a URI, as in

    proxy_pass http://127.0.0.1/url/;

then when a request is passed to the server, the request URI matching
the location is replaced by the URI in the proxy_pass directive. This is
what wiped the query parameters.

If proxy_pass is not specified with a URI, as in

    proxy_pass http://127.0.0.1;

then the request URI is passed to the server in the same form as
received from the client (with all query parameters).

This, of course, relies on the research API container serving the right
thing on the /generate location, but that happens to be the case for me.

The other solution would have been to parse the query parameters from
the $request_uri.

The following Bash snippet was taken from a script I've been using to
nicely query the API, and funnel the returned JSON through jq to make it
more human readable.

    # The quoting and array expansion here was a bitch to get right.
    CURL_PROMPT=()
    if [ -n "$PROMPT" ]; then
        CURL_PROMPT=(--data-urlencode "prompt=$PROMPT")
    fi
    CURL_SEED=()
    if [ -n "$SEED" ]; then
        CURL_SEED=(--data-urlencode "seed=$SEED")
    fi

    curl \
        --silent \
        --show-error \
        --get \
        --write-out "%{stderr}%{url_effective}\n" \
        "${CURL_PROMPT[@]}" \
        "${CURL_SEED[@]}" \
        --data-urlencode "number=$NUMBER" \
        --data-urlencode "temperature=$TEMPERATURE" \
        --data-urlencode "k=0" \
        --data-urlencode "p=0.9" \
        --data-urlencode "max_tokens=$MAX_TOKENS" \
        "https://agill.xyz/generate"

It's been handy for testing the generation API.

[1] 5779ffda83a3b517a3f3e2588f6be3210a7d1902
[2] https://nginx.org/en/docs/http/ngx_http_proxy_module.html#proxy_pass